feat: 부서 CRUD 기능 구현

+++ src/main/java/com/kooyeoung/hrbank/dto/repository/DepartmentSearchCondition.java
@@ -4,7 +4,7 @@ public record DepartmentSearchCondition(
         String keyword
         ,String sortFiled
         ,String cursor
-        ,int idAfter
+        ,long idAfter
         , boolean hasCursor
         , boolean isDesc
         , int size
